+++ apps/webapp/app/presenters/v3/RunListPresenter.server.ts
@@ -168,6 +168,7 @@ export class RunListPresenter extends BasePresenter {
         usageDurationMs: BigInt;
         tags: string[];
         depth: number;
+        rootTaskRunId: string | null;
       }[]
     >`
     SELECT
@@ -192,6 +193,7 @@ export class RunListPresenter extends BasePresenter {
     tr."costInCents" AS "costInCents",
     tr."usageDurationMs" AS "usageDurationMs",
     tr."depth" AS "depth",
+    tr."rootTaskRunId" AS "rootTaskRunId",
     array_remove(array_agg(tag.name), NULL) AS "tags"
 FROM
     ${sqlDatabaseSchema}."TaskRun" tr
@@ -336,6 +338,7 @@ WHERE
           usageDurationMs: Number(run.usageDurationMs),
           tags: run.tags.sort((a, b) => a.localeCompare(b)),
           depth: run.depth,
+          rootTaskRunId: run.rootTaskRunId,
         };
       }),
       pagination: {