+++ apps/webapp/app/presenters/v3/SpanPresenter.server.ts
@@ -1,4 +1,9 @@
-import { MachinePresetName, prettyPrintPacket, TaskRunError } from "@trigger.dev/core/v3";
+import {
+  MachinePresetName,
+  parsePacket,
+  prettyPrintPacket,
+  TaskRunError,
+} from "@trigger.dev/core/v3";
 import { RUNNING_STATUSES } from "~/components/runs/v3/TaskRunStatus";
 import { eventRepository } from "~/v3/eventRepository.server";
 import { machinePresetFromName } from "~/v3/machinePresets.server";
@@ -113,6 +118,8 @@ export class SpanPresenter extends BasePresenter {
         },
         payload: true,
         payloadType: true,
+        metadata: true,
+        metadataType: true,
         maxAttempts: true,
         project: {
           include: {
@@ -123,6 +130,21 @@ export class SpanPresenter extends BasePresenter {
           select: {
             filePath: true,
             exportName: true,
+          },
+        },
+        //relationships
+        rootTaskRun: {
+          select: {
+            taskIdentifier: true,
+            friendlyId: true,
+            spanId: true,
+          },
+        },
+        parentTaskRun: {
+          select: {
+            taskIdentifier: true,
+            friendlyId: true,
+            spanId: true,
           },
         },
       },
@@ -184,6 +206,10 @@ export class SpanPresenter extends BasePresenter {
     }
 
     const span = await eventRepository.getSpan(spanId, run.traceId);
+
+    const metadata = run.metadata
+      ? await prettyPrintPacket(run.metadata, run.metadataType)
+      : undefined;
 
     const context = {
       task: {
@@ -270,8 +296,17 @@ export class SpanPresenter extends BasePresenter {
       output,
       outputType: finishedAttempt?.outputType ?? "application/json",
       error,
-      links: span?.links,
+      relationships: {
+        root: run.rootTaskRun
+          ? {
+              ...run.rootTaskRun,
+              isParent: run.parentTaskRun?.friendlyId === run.rootTaskRun.friendlyId,
+            }
+          : undefined,
+        parent: run.parentTaskRun ?? undefined,
+      },
       context: JSON.stringify(context, null, 2),
+      metadata,
     };
   }
 
@@ -295,10 +330,29 @@ export class SpanPresenter extends BasePresenter {
       return;
     }
 
+    const triggeredRuns = await this._replica.taskRun.findMany({
+      select: {
+        friendlyId: true,
+        taskIdentifier: true,
+        spanId: true,
+        createdAt: true,
+        number: true,
+        lockedToVersion: {
+          select: {
+            version: true,
+          },
+        },
+      },
+      where: {
+        parentSpanId: spanId,
+      },
+    });
+
     return {
       ...span,
       events: span.events,
       properties: span.properties ? JSON.stringify(span.properties, null, 2) : undefined,
+      triggeredRuns,
       showActionBar: span.show?.actions === true,
     };
   }