+++ packages/core/src/v3/run-metadata-api.ts
@@ -1,0 +1,5 @@
+// Split module-level variable definition into separate files to allow
+// tree-shaking on each api instance.
+import { RunMetadataAPI } from "./runMetadata/index.js";
+
+export const runMetadata = RunMetadataAPI.getInstance();